+++ accounts/forms.py
@@ -1,0 +1,7 @@
+from django import forms
+from django.contrib.auth.models import User
+
+class RegisterForm(forms.ModelForm):
+    class Meta:
+        model = User
+        fields = ['password','username','email']
