+++ accounts/models.py
@@ -1,7 +1,2 @@
 from django.db import models
 from django.contrib.auth.models import User
-
-# Create your models here.
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    bio = models.TextField(blank=True)
