+++ accounts/templates/register.html
@@ -9,4 +9,5 @@
         {{ form }}
         <button type="submit" value ="submit">Submit</button>
     </form>
+    
 {% endblock %}
