+++ accounts/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from .views import register
+app_name = "accounts"
+
+urlpatterns = [
+    path('register', register, name='register'),
+
+]
