+++ user_management/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.contrib.auth.models import User
 from .models import Profile
-#test
+
 class ProfileInline(admin.StackedInline):
     model = Profile
     can_delete = False
