+++ user_management/forms.py
@@ -1,5 +1,4 @@
 from django import forms
-from django.contrib.auth.models import User
 
 from .models import Profile
 
@@ -7,8 +6,3 @@ class ProfileForm(forms.ModelForm):
     class Meta:
         model = Profile
         fields = ['display_name', 'email_address']
-
-class RegisterForm(forms.ModelForm):
-    class Meta:
-        model = User
-        fields = ['password','username','email']
