+++ user_management/views.py
@@ -2,27 +2,8 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 
-from .forms import ProfileForm, RegisterForm
+from .forms import ProfileForm
 from .models import Profile
-
-def register(request):
-    form = RegisterForm()
-    if request.method == 'POST':
-        form = RegisterForm(request.POST)
-    if form.is_valid():
-        user = form.save(commit=False)
-        user.set_password(form.cleaned_data['password']) 
-        user.save()
-
-        Profile.objects.create(
-                user=user,
-                display_name=user.username,
-                email_address=user.email  
-            )
-        
-        return redirect('login')
-    ctx = {"form": form }
-    return render(request, 'register.html', ctx)
 
 @login_required
 def profile_update(request, username):
